+++ modules/AWSPowerShell/Cmdlets/AmazonEC2/Basic/Edit-EC2IdFormat-Cmdlet.cs
@@ -28,12 +28,12 @@ using Amazon.EC2.Model;
 namespace Amazon.PowerShell.Cmdlets.EC2
 {
     /// <summary>
-    /// <b>Important: This command is reserved for future use, and is currently not available
-    /// for you to use.</b><para>
     /// Modifies the ID format for the specified resource on a per-region basis. You can specify
     /// that resources should receive longer IDs (17-character IDs) when they are created.
     /// The following resource types support longer IDs: <code>instance</code> | <code>reservation</code>.
-    /// </para><para>
+    /// 
+    ///  
+    /// <para>
     /// This setting applies to the IAM user who makes the request; it does not apply to the
     /// entire AWS account. By default, an IAM user defaults to the same settings as the root
     /// user, unless they explicitly override the settings by running this request. Resources